+++ Gruntfile.js
@@ -39,6 +39,9 @@ module.exports = function(grunt) {
       common: {
         src: ['common/**/*.js']
       },
+      browser: {
+        src: ['browser/**/*.js']
+      },
       server: {
         src: ['server/**/*.js']
       },
@@ -51,6 +54,7 @@ module.exports = function(grunt) {
       lib: ['lib/**/*.js'],
       common: ['common/**/*.js'],
       server: ['server/**/*.js'],
+      browser: ['browser/**/*.js'],
       test: ['test/**/*.js']
     },
     watch: {
