+++ docs.json
@@ -5,6 +5,7 @@
     "lib/server-app.js",
     "lib/loopback.js",
     "lib/registry.js",
+    "server/current-context.js",
     "lib/access-context.js",
 { "title": "Base models", "depth": 2 },
     "lib/model.js",
@@ -13,7 +14,7 @@
     "server/middleware/context.js",
     "server/middleware/favicon.js",
     "server/middleware/rest.js",
-    "server/middleware/static.js",    
+    "server/middleware/static.js",
     "server/middleware/status.js",
     "server/middleware/token.js",
     "server/middleware/url-not-found.js",
@@ -21,11 +22,11 @@
     "common/models/access-token.js",
     "common/models/acl.js",
     "common/models/application.js",
-    "common/models/change.js",    
+    "common/models/change.js",
     "common/models/email.js",
     "common/models/role.js",
     "common/models/role-mapping.js",
-    "common/models/scope.js",    
+    "common/models/scope.js",
     "common/models/user.js"
   ],
   "assets": "/docs/assets"
