+++ lib/loopback.js
@@ -189,10 +189,7 @@ loopback.template = function(file) {
   return ejs.compile(str);
 };
 
-loopback.getCurrentContext = function() {
-  // A placeholder method, see lib/middleware/context.js for the real version
-  return null;
-};
+require('../server/current-context')(loopback);
 
 /*!
  * Built in models / services
